Update monkey patches
